Created basic listing framework.
Still need to work out passing parameters to activities
before I can call it closed. Currently all of the target
entity are displayed rather than filtered by the selection.

+++ app/src/main/java/com/forthecoder/collegeschedule/entity/BaseEntity.java
@@ -2,13 +2,13 @@ package com.forthecoder.collegeschedule.entity;
 
 class BaseEntity {
 
-    private Integer rowid;
+    private Long rowid;
 
-    public Integer getRowid() {
+    public Long getRowid() {
         return rowid;
     }
 
-    public void setRowid(Integer rowid) {
+    public void setRowid(Long rowid) {
         this.rowid = rowid;
     }
 }
